requirements.txt
Flask_RESTful==0.3.7
celery==4.2.1
docopt==0.6.2
Flask_Script==2.0.6
Flask==1.0.2
requests==2.18.4
redis==3.3.11
jmespath==0.9.3
Flask_Cors==3.0.7
python_dateutil==2.8.1
python-dotenv==0.11.0
elasticsearch==7.5.1
flask_swagger_ui==3.20.9
pymysql==0.9.3
selenium==3.141.0
sshtunnel==0.1.5
PyYAML==5.3
